Move style modules
