+++ frontend/components/common/PokemonDescription.module.scss
@@ -1,4 +1,4 @@
-@use 'customVariables' as c-vars;
+@use '../../styles/customVariables' as c-vars;
 @use '@carbon/styles/scss/theme';
 @use '@carbon/layout';
