+++ frontend/components/common/Popularity.module.scss
