+++ frontend/components/common/QuickViewModal.module.scss
